Move renderTabFn to Navigation.js
