+++ app/containers/Stage.js
@@ -18,21 +18,11 @@ import VipCodeEntry           from './VipCodeEntry'
 import Quiz                   from './Quiz'
 import WaitingRoom            from './WaitingRoom'
 import Dead                   from './Dead'
-import Icon                   from 'react-native-vector-icons/Ionicons'
 import {em}                   from '../constants/dimensions'
 import timing                 from '../constants/timing'
 import {
-  mayteWhite,
-  mayteBlack,
-} from '../constants/colors'
-import {
   StyleSheet,
-  Text,
   View,
-  ScrollView,
-  FlatList,
-  Animated,
-  TouchableWithoutFeedback,
 } from 'react-native'
 
 const useScratch = false
@@ -75,17 +65,16 @@ class Stage extends PureComponent {
     :
       <Navigation initialSceneName="VelvetRope">
         <Settings sceneName="Settings"
-          tabLabel={renderTabFn({label: 'Settings', iconFocused: 'ios-person', iconUnfocused: 'ios-person-outline'})}/>
-        <VelvetRope
-          sceneName="VelvetRope"
-          tabLabel={renderTabFn({label: 'Membership', iconFocused: 'ios-key', iconUnfocused: 'ios-key-outline'})}/>
+          label="Settings" iconFocused="ios-person" iconUnfocused="ios-person-outline" />
+        <VelvetRope sceneName="VelvetRope"
+          label="Membership" iconFocused="ios-key" iconUnfocused="ios-key-outline"/>
 
         { props.isAdmin ?
           <Recs sceneName="Recs"
-            tabLabel={renderTabFn({label: 'Suggestions', iconFocused: 'ios-heart', iconUnfocused: 'ios-heart-outline'})}/>
+            label="Suggestions" iconFocused="ios-heart" iconUnfocused="ios-heart-outline" />
         :
-        <RecsPreview sceneName="Recs"
-          tabLabel={renderTabFn({label: 'Suggestions', iconFocused: 'ios-heart', iconUnfocused: 'ios-heart-outline'})}/>
+          <RecsPreview sceneName="Recs"
+            label="Suggestions" iconFocused="ios-heart" iconUnfocused="ios-heart-outline"/>
         }
       </Navigation>
     )
@@ -123,61 +112,8 @@ function mapDispatchToProps(dispatch) {
   }
 }
 
-const renderTabFn = (props) => {
-  let initRender = false
-  let _focusOp = new Animated.Value(0)
-  let _unfocusOp = new Animated.Value(0)
-  return ({tintColor, focused}) => {
-    if (!initRender) {
-      initRender = true
-      _focusOp = focused ? new Animated.Value(1) : new Animated.Value(0)
-      _unfocusOp = focused ? new Animated.Value(0) : new Animated.Value(1)
-    }
-
-    if (initRender && focused) {
-      Animated.parallel([
-        Animated.timing(_focusOp, {toValue: 1, duration: timing.navFade, useNativeDriver: true}),
-        Animated.timing(_unfocusOp, {toValue: 0, duration: timing.navFade, useNativeDriver: true}),
-      ]).start()
-    } else {
-      Animated.parallel([
-        Animated.timing(_focusOp, {toValue: 0, duration: timing.navFade, useNativeDriver: true}),
-        Animated.timing(_unfocusOp, {toValue: 1, duration: timing.navFade, useNativeDriver: true}),
-      ]).start()
-    }
-
-    const isFocused = tintColor == 'gainsboro';
-    return (
-      <View style={style.navTabCont}>
-        <Animated.View style={[style.navTabCont, {opacity: _focusOp}]}>
-          <LinearGradient colors={['rgb(68,64,65)', mayteBlack()]} style={[style.navTabGradient]}></LinearGradient>
-          <Icon name={props.iconFocused}
-                style={[style.navTabIcon, {color: tintColor}]}
-                size={em(1.625)} />
-          <Text style={[style.navTabLabel, {color: mayteWhite()}]}>
-            {props.label}
-          </Text>
-        </Animated.View>
-        <Animated.View style={[style.navTabCont, {opacity: _unfocusOp}]}>
-          <LinearGradient colors={[mayteWhite(), 'rgb(225,221,222)']} style={[style.navTabGradient]}></LinearGradient>
-          <Icon name={props.iconUnfocused}
-                style={[style.navTabIcon, {color: tintColor}]}
-                size={em(1.625)} />
-          <Text style={[style.navTabLabel, {color: mayteBlack()}]}>
-            {props.label}
-          </Text>
-        </Animated.View>
-      </View>
-    )
-  }
-}
-
 const style = StyleSheet.create({
   container: {flex: 1,},
-  navTabCont: {position: 'absolute', top: 0, left: 0, width: '100%', height: '100%', justifyContent: 'center', alignItems: 'center',},
-  navTabIcon: {backgroundColor: 'transparent', marginTop: em(0.4)},
-  navTabGradient: {position: 'absolute', top: 0, left: 0, bottom: 0, right: 0},
-  navTabLabel: {fontFamily: 'Gotham-Book', letterSpacing: em(0.1), fontSize: em(0.5), marginBottom: em(0.33), backgroundColor: 'transparent'},
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(Stage)
